refactor : apply decorator and seperate utility class
